Pre-emptively declare some globals in windows.py
The current implementation relies on a call to init() to *declare( the global
variables.  Pyflakes doesn't / can't know this, so it assumes the usage of these
variables in the other functions in this module is erroneous.

To alleviate the warnings, declare these variables as `None` in the
module scope (where they end up anyway)